Use runtime metrics for container memory metrics
Most "files" in the virtual file system `/proc/` return metrics for the
host system rather than the container in which it is fetched. To get the
values for the host metrics in containers we need to read the `/sys/fs/`
virtual file system instead.

This commit fixes the memory metrics reporting for containers that
expose the `/sys/fs/` virtual file system.

Note: This change does hard code the memory reporting to containers,
this needs to be automatically detected in the future.

Based on the docs as described here:
https://docs.docker.com/config/containers/runmetrics/

Something in the calculations for `/proc/meminfo` was messing up the
memory calculations for container metrics. This change moves the
calculations to the read functions so the implementation of the
calculations can differ for these two reporting methods.

To get accurate "used" memory, it was needed to subtract page cache from
used memory. This also allows us to get the free memory.

As implemented in `docker stats`
https://github.com/docker/cli/blob/1401d5daf2f49a97791487dd5c5a8598907f0bf1/cli/command/container/stats_helpers.go#L226-L229