Use runtime metrics for container memory metrics
Most "files" in the virtual file system `/proc/` return metrics for the
host system rather than the container in which it is fetched. To get the
values for the host metrics in containers we need to read the `/sys/fs/`
virtual file system instead.

This commit fixes the memory metrics reporting for containers that
expose the `/sys/fs/` virtual file system.

Note: This change does hard code the memory reporting to containers,
this needs to be automatically detected in the future.

Based on the docs as described here:
https://docs.docker.com/config/containers/runmetrics/