Use runtime metrics for container CPU metrics
Most "files" in the virtual file system `/proc/` return metrics for the
host system rather than the container in which it is fetched. To get the
values for the host metrics in containers we need to read the `/sys/fs/`
virtual file system instead.

This commit fixes the CPU metrics reporting for containers that
expose the `/sys/fs/` virtual file system.

Since only `user`, `system` and `total` are exposed in these runtime
metrics a lot of values default to 0. It might be interesting to avoid
reporting these as they don't have actual values. The value `None`,
which would mean wrapping it in an `Option`, would maybe work better
here.

Based on the docs as described here:
https://docs.docker.com/config/containers/runmetrics/