Fix swap reporting
All values for swap have the memory values included, so subtract the
memory total and usage to get the swap values.

Also don't use option for used swap. It's not needed, just set to 0 if
not found.

Read from the dedicated files as `memory.stat` will report `swap: 0` a
lot of the times. It doesn't seem as reliable as the dedicated files.