Fix container detection
Don't rely on cgroup names, but rather check the existence of the
virtual file system reporting the container metrics. If it exists, we
can assume it contains the metrics we need.

An advantage of this method is that we don't have to keep a exhaustive
list of container cgroup names and patterns to match against.

As previously mentioned here:
https://github.com/appsignal/probes-rs/pull/27#issuecomment-392917623

This also fixes lack of memory reporting for hosts that are identified
as containers, but don't have this virtual file system mounted. They
will be detected as "normal" machines and read from `/proc`. This is the
current behavior in the AppSignal agent and we don't want to change that
right now.