Fix time_adjusted calculations (#40)
## Some context about the `time_adjusted` function

The `time_adjusted` function returns a value calculated for a minute. It
received two values, for which it calculates the delta first. Then,
using the `time_difference_ns` value (the time in which these values
were measured), it calculates the value as it would be for a full
minute.

When the `time_difference_ns` is 59 seconds or 61 seconds it returns the
value as it would be for 60 seconds instead. This helps account for
small drift within the time at which the metrics were measured.

## The problem

The previous implementation of the time_adjusted function would multiple
the delta value by the real time (30 seconds) and divide that value by
60 seconds.

The problem here was that it would not return the value for 60 seconds,
but instead return that for around 15 seconds, half of 30 seconds.

```
delta = 10.0
time_difference = 30.0
expected_time = 60.0

delta * real_time / expected_time
=> 5.0
```

5.0 being half of 10.0 (the delta), half of 30 seconds.

## The fix

Instead of multiplying the delta by `time_difference_ns` we should
divide the delta by the `time_difference_ns` to get the value per
nanosecond (ns). Then we multiply that 1 nanosecond value by 60 seconds.
This will return the average value for 60 seconds.

```
delta = 10.0
time_difference = 30.0
expected_time = 60.0

delta / real_time * expected_time
=> 20.0
```

20.0 being double of 10.0 (the delta), double of 30 seconds.