Zero Rust warnings
Go down to zero warnings for Rust 1.43 and adjust formatting to comply
with current rustfmt default settings. This change does not have much
value in its own. But having more idiomatic Rust code makes it easier
for others to contribute and minizies diffs in the future.

Minimum supported Rust version is now 1.36 due to the use of
std::mem::MaybeUninit.

Some of the things done:

- Fix redundant struct initializers
- Fix redundant imports
- Fix unnecessary to_owned() calls
- Fix single-char str patterns
- Use ok_or_else() to avoid unnecessary code execution on the unhappy
  path
- Fix repo url in Cargo.toml

Also reformat the whole source with `cargo fmt` in its default settings.