Report shared memory metric (#59)
Report the shared memory metric. This supports it for VMs, and
containers using cgroups v1. Cgroups v2 support is tracked in issue #57.

Remove the required fields check for the `/sys/fs` system, because the
`shmem` metric is not available in every version of the
`/sys/fs/cgroup/memory/memory.stat` file we've seen.
This is less efficient, because it now needs to parse the entire file,
but it's optional field, so we have to.

Part of https://github.com/appsignal/support/issues/227

Co-authored-by: Luismi Ramírez <luismi.ramirez@protonmail.com>