Install the right Rust toolchain on build
Instead of installing the default toolchain (latest stable) when
installing `rustup` during the build, then having the correct
toolchain (as specified by `.rust-toolchain`) installed when the
tests are ran, install `rustup` without a toolchain and then ask
it to install the toolchain in `.rust-toolchain`.

Do not remove `Cargo.lock` from the test container, as its absence
causes Cargo to update its index in order to rebuild it.