Support Alpine Linux disk usage reporting using df
On Ubuntu, `df` lists all disk by default. Using the `-l/--local`
flag, it only lists the local disks, as per the help output:

```
-l, --local           limit listing to local file systems
```

On Alpine Linux the `-l/--local` flag does not exist. Instead it has a
`-a` flag to do the opposite:

```
-a	Show all filesystems
```

Try running `df` without any flags when `df --local` doesn't work. (Use
the full `--local` flag name so it's easier to figure out what it's
doing when reading the code.)
This way we can parse the output on Alpine Linux.