Omit inodes for mountpoints with a dash percentage (#73)
We had a report from someone with `df -i` output that reported a dash
(-) for the inodes used percentage for a mountpoint.

I can only reproduce this with a `df -i --all` call, but it's good to
safeguard against the percentage being a dash. When no percentage is
found, skip the mountpoint entirely.

We could also make the `iused_percentage` field an Option, but this is a
smaller change, fixing a possible error. This won't change the public
API. Our agent only uses the percentage, so making it an Option wouldn't
make a difference for our metrics reporting.

Closes #68