Add support for Aptible CPU share env var
Use the Aptible CPU shares env var to determine the number of CPUs on
the Aptible hosting provider, and so normalize CPU percentages for apps
running on Aptible.

> APTIBLE_CONTAINER_CPU_SHARE
> Provides the vCPU share for the container, matching the ratios in our
> documentation for container profiles. Format will be provided in the
> following format: 0.125, 0.5, 1.0, etc.

https://www.aptible.com/docs/aptible-metadata-variables

Because the tests set and unset an env var, the Rust test threads need
to be set to 1, otherwise the env var leaks into other tests.